Created code generator
